Update Average Pooling layer
The code is not right on average pooling layer with odd input image size.  For example, input image with 57x57, kernel size with 2x2 and stride 2,  will generate wrong results. The updated code would fix this corner case, but the performance may be affected by this update code which will be left for further optimization.